rework purpose of fido_cred_set_type
rework purpose of fido_cred_set_type after feedback in issue #6:
issue as a hint of what the expected credential type should be, and
match it against the type obtained from the authenticator. when
verifying, use it as a hint when parsing authdata, so that
subsequent calls to fido_cred_pubkey_{ptr,len} return the expected
length and type.

add a regression test while here.